backend/requirements.txt
Django>=3.0,<4.0
celery==5.1.2
redis==3.5.3
gunicorn==20.1.0
psycopg2-binary>=2.8
django-cors-headers
djangorestframework
PyJWT
django-extensions
requests
django-crispy-forms
whitenoise
